cmdio: relocate helpers into dedicated files; build Secret/Select on Run*
compat.go is gone. Log/LogString move to log.go, and
readLine/Ask/AskYesOrNo move to ask.go. The "compatibility layer" doc
comments are replaced with descriptions of what each function does,
since these are now their permanent home.

Tuple/Select/SelectOrdered move from io.go to select.go and are
reimplemented on top of RunSelect. A new HideSelected option on
SelectOptions preserves the existing post-prompt display behavior.

Secret moves from io.go to prompt.go and becomes a thin wrapper over
RunPrompt. A new HideEntered option on PromptOptions carries the
masked-input post-submission behavior.
