+++ libs/cmdio/prompt.go
@@ -18,6 +18,9 @@ type PromptOptions struct {
 	// (use '*' for password-style input).
 	Mask rune
 
+	// HideEntered hides the entered value after the prompt closes.
+	HideEntered bool
+
 	// Validate, when set, is called on every keystroke; returning a non-nil
 	// error keeps the prompt open and shows the error to the user.
 	Validate func(input string) error
@@ -27,12 +30,23 @@ type PromptOptions struct {
 func RunPrompt(ctx context.Context, opts PromptOptions) (string, error) {
 	c := fromContext(ctx)
 	p := promptui.Prompt{
-		Label:    opts.Label,
-		Default:  opts.Default,
-		Mask:     opts.Mask,
-		Validate: opts.Validate,
-		Stdin:    c.promptStdin(),
-		Stdout:   nopWriteCloser{c.err},
+		Label:       opts.Label,
+		Default:     opts.Default,
+		Mask:        opts.Mask,
+		HideEntered: opts.HideEntered,
+		Validate:    opts.Validate,
+		Stdin:       c.promptStdin(),
+		Stdout:      nopWriteCloser{c.err},
 	}
 	return p.Run()
+}
+
+// Secret prompts the user for a value while masking input with '*' and hiding
+// the entered value after submission.
+func Secret(ctx context.Context, label string) (string, error) {
+	return RunPrompt(ctx, PromptOptions{
+		Label:       label,
+		Mask:        '*',
+		HideEntered: true,
+	})
 }
